### 2019-07-11 #1565 @whedon
[PRE REVIEW]: kdensity: An R package for kernel density estimation with parametric starts and asymmetric kernels
**Submitting author:** @JonasMoss (<a href="http://orcid.org/0000-0002-6876-6964">Jonas Moss</a>)
**Repository:** <a href="https://github.com/JonasMoss/kdensity" target ="_blank">https://github.com/JonasMoss/kdensity</a>
**Version:** v1.0.1
