### 2019-07-11 #1565 @arfon
@JonasMoss - I don't think Pandoc supports `\code`. Can you use back-ticks instead?
