### 2019-07-11 #1565 @terrytangyuan
@arfon Yes definitely.

@JonasMoss if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, this list of people have already agreed to review for JOSS and may be suitable for this submission.
