+++ paper/paper.bib
@@ -4,7 +4,7 @@
   journal={The Annals of Statistics},
   pages={882--904},
   year={1995},
-  publisher={JSTOR}
+  publisher={JSTOR},
   doi = {doi:10.1214/aos/1176324627}
 }
 
@@ -16,7 +16,7 @@
   number={4},
   pages={977--984},
   year={2007},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi = {doi:10.1093/biomet/asm068}
 }
 
@@ -28,7 +28,7 @@
   number={3},
   pages={471--480},
   year={2000},
-  publisher={Springer}
+  publisher={Springer},
   doi = {doi:10.1023/A:1004165218295}
 }
 
@@ -40,7 +40,7 @@
   number={2},
   pages={131--145},
   year={1999},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi = {doi:10.1016/S0167-9473(99)00010-9}
 }
 
@@ -48,7 +48,7 @@
   title={Density estimation for statistics and data analysis},
   author={Silverman, Bernard W},
   year={2018},
-  publisher={Routledge}
+  publisher={Routledge},
   doi = {doi:10.1201/9781315140919}
 }
 
@@ -60,7 +60,7 @@
   number={4},
   pages={653--671},
   year={1994},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi = {doi:10.2307/2346189}
 }
 
@@ -70,5 +70,5 @@
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
     year = {2019},
-    url = {https://www.R-project.org/},
+    url = {https://www.R-project.org/}
   }
