### 2019-07-12 #1565 @JonasMoss
Changed \code to backticks. I suggest @TrashBirdEcology for reviewing.
