### 2019-07-12 #1566 @terrytangyuan
@TrashBirdEcology I also added you as the reviewer since JOSS now requires at least two reviewers. Feel free to review this paper when you get back. Thanks.
