### 2019-07-22 #1566 @terrytangyuan
👋 @gvegayon You mentioned that you would be able to start the review on 22nd. Feel free to start the review while @TrashBirdEcology is still traveling.
