### 2019-07-24 JonasMoss/kdensity#49 @gvegayon
[JOSS Review] summary method
Part of JOSS review openjournals/joss-reviews#1566

This is more style than anything, but I think it would be nice it the computed quantities were returned instead of actual object in `summary.kdensity`.
