### 2019-07-24 JonasMoss/kdensity#50 @gvegayon
[JOSS Review] Vignette
Part of JOSS review openjournals/joss-reviews#1566

This is more a question than anything, how this statement is true?

https://github.com/JonasMoss/kdensity/blob/ade1d128517b106364a071b29e85af4e0aee8e44/vignettes/tutorial.Rmd#L137

I mean, I was expecting to see a big difference (improvement) on the curve fitting in order to say that, but it seems that the normal kd is doing a fair job (perhaps better?). This from the fact that I'm not an expert in parametric/nonparametric kernel models. It would be nice to have a better explanation of this in the vignette :).
