### 2019-07-24 JonasMoss/kdensity#51 @gvegayon
[JOSS Review] Related R packages
Part of JOSS review openjournals/joss-reviews#1566

I think it is important to list and point out differences with other R packages that may perform a similar task, e.g.:

- [kdcopula](https://cran.r-project.org/package=kdecopula/) for copula bivariate kernel density.
- [kde1d](https://cran.r-packages.org/package=kde1d) for bounded univariate kernel density.
- [Ake](https://cran.r-packages.org/package=Ake) and [ks](https://cran.r-packages.org/package=ks) which implement the Beta kernel of Chen (1999).
- etc.

This should be somehow reflected in both the README.(R)md file and the paper.md file.
