### 2019-07-24 JonasMoss/kdensity#52 @gvegayon
[JOSS Review] Contribution/bug report guidelines
Part of JOSS review openjournals/joss-reviews#1566

The package should have explicit guidelines for contributing to the project/asking questions/reporting bugs. See https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines
