### 2019-07-24 #1566 @gvegayon
OK @JonasMoss @terrytangyuan, I've finished my first round. Overall, everything looks good (very nice job @JonasMoss!), and there are only a few issues that I would like the authors to address. The most important is regarding related R packages.
