### 2019-07-31 #1566 @terrytangyuan
Thanks @gvegayon! @JonasMoss Could you address those issues pointed out?
