### 2019-08-04 46d9d66
:books: Response to [JOSS Review] Documentation.
Updated documentation of lines and points. Added restrictions to the $ and [[]] operators, making them compliant with the documentation. Some other changes made to pass Rcmd check and testthat tests.
