### 2019-08-04 JonasMoss/kdensity#48 @JonasMoss
I've responded to the two first points in 46d9d6670380e3d09030b179801857e123213ac0. I can't document all the members of the environment since some are internal, but I restricted the elements [[]] and $ can access. Thanks for spotting this!

I'm not sure about enumerating all the bandwidth selectors since it would overpopulate the documentation. (I prefer it lean!) And it would pressure me to enumerate the documentation for all the starts and kernels too, further overpopulating the documentation.
