### 2019-08-04 JonasMoss/kdensity#49 @JonasMoss
The computed quantities are all available in the object itself. The summary function only works as an extended print, it does no computations itself. E.g. the bandwidth is available in object$bw_str etc.
