### 2019-08-04 e5148f4
:book: [JOSS Review] Vignette
Added an explantion for why we are justified in using the skew hyperbolic t.
