### 2019-08-04 JonasMoss/kdensity#50 @JonasMoss
Made a change in e5148f40275f3799b6fe26f7ea66706e78b235ee. My only point is that you don't appear to need KDE in order to have a good fit to the density. Since KDEs are more complex and parametric densities, I think it's better to use parametrics when you can.

So no hidden expertise in non-parametric models involved here, sadly. :p
