### 2019-08-04 9acec1b
:book: [JOSS Review] Related R packages
Added references to R packages doing similar stuff.
