### 2019-08-04 JonasMoss/kdensity#51 @JonasMoss
Updated readme in 9acec1b707e91fc521437aa4d1f6ecacda2cb0a8. Will look at the paper too.
