### 2019-08-04 b98b0ae
:book: [JOSS Review] Contribution/bug report guidelines.
Added contribution guidelines and code of conduct.
