+++ README.Rmd
@@ -111,6 +111,12 @@ coef(kde)
 logLik(kde)
 AIC(kde)
 ```
+## How to Contribute or Get Help
+If you encounter a bug, have a feature request or need some help, don't hesitate
+to open an [issue](https://github.com/JonasMoss/kdensity/issues). If you want to
+contribute, make a pull request. This project follows a 
+[Contributor Code of Conduct](https://www.contributor-covenant.org/version/1/4/code-of-conduct.md).
+
 ## References
 
 * [Hjort, Nils Lid, and Ingrid K. Glad. "Nonparametric density estimation with a parametric start." The Annals of Statistics (1995): 882-904.](https://projecteuclid.org/euclid.aos/1176324627).
