### 2019-08-04 JonasMoss/kdensity#52 @JonasMoss
Thanks! I added some guidelines in b98b0aecc1143260894b87120fbc7419faef7d74.
