### 2019-08-05 c684555
:newspaper: [JOSS Review] Related R packages
Added references to other work on KDEs in R.
