+++ paper/paper.bib
@@ -72,3 +72,56 @@
     year = {2019},
     url = {https://www.R-project.org/}
   }
+  
+@article{deng2011,
+	title = {Density estimation in R},
+	author = {Deng, Henry and Wickham, Hadley},
+	year = {2011},
+	url = {http://vita.had.co.nz/papers/density-estimation.pdf}
+}
+
+@Article{evmix,
+	title = {{evmix}: An {R} package for Extreme Value Mixture Modeling, Threshold Estimation and Boundary Corrected Kernel Density Estimation},
+    author = {Yang Hu and Carl Scarrott},
+    journal = {Journal of Statistical Software},
+    year = {2018},
+    volume = {84},
+    number = {5},
+    pages = {1--27},
+    doi = {10.18637/jss.v084.i05},
+	url = {http://www.math.canterbury.ac.nz/~c.scarrott/evmix/}
+}
+
+@Manual{kde1d,
+    title = {kde1d: Univariate Kernel Density Estimation},
+    author = {Thomas Nagler and Thibault Vatter},
+    year = {2019},
+    note = {R package version 0.4.0},
+    url = {https://CRAN.R-project.org/package=kde1d},
+}
+
+@article{logKDE,
+  title={logKDE: log-transformed kernel density estimation.},
+  author={Jones, Andrew T and Nguyen, Hien Duy and McLachlan, Geoffrey J},
+  journal={J. Open Source Software},
+  volume={3},
+  number={28},
+  pages={870},
+  year={2018}
+}
+
+@Manual{ks,
+    title = {ks: Kernel Smoothing},
+    author = {Tarn Duong},
+    year = {2019},
+    note = {R package version 1.11.5},
+    url = {https://CRAN.R-project.org/package=ks},
+}
+
+@Manual{Ake,
+    title = {Ake: Associated Kernel Estimations},
+    author = {W. E. Wansouwé and S. M. Somé and C. C. Kokonendji},
+    year = {2015},
+    note = {R package version 1.0},
+    url = {https://CRAN.R-project.org/package=Ake},
+}
