### 2019-08-05 0878c0e
[JOSS Review] Related R packages
Added DOI; minor change to paper.
