+++ paper/paper.bib
@@ -101,13 +101,14 @@
 }
 
 @article{logKDE,
-  title={logKDE: log-transformed kernel density estimation.},
-  author={Jones, Andrew T and Nguyen, Hien Duy and McLachlan, Geoffrey J},
-  journal={J. Open Source Software},
-  volume={3},
-  number={28},
-  pages={870},
-  year={2018}
+  title = {logKDE: log-transformed kernel density estimation.},
+  author = {Jones, Andrew T and Nguyen, Hien Duy and McLachlan, Geoffrey J},
+  journal = {Journal of Open Source Software},
+  volume = {3},
+  number = {28},
+  pages = {870},
+  year = {2018},
+  doi = {10.21105/joss.00870}
 }
 
 @Manual{ks,
