### 2019-08-05 JonasMoss/kdensity#51 @JonasMoss
Updated the paper too. I think think the main difference between this package and all others is that this one uses parametric starts.
