### 2019-08-05 #1566 @JonasMoss
@gvegayon Okay! I've addressed the issues. Details are on the Github issues pages.
