### 2019-08-06 JonasMoss/kdensity#56 @trashbirdecology
[JOSS Review]: Installation instructions: missing dependency: EQL
JOSS Review [1566](https://github.com/openjournals/joss-reviews/issues/1566)

Either add as a depends or specify the dependency in the README for package `EQL`

`Error: The bandwidth function 'RHE' requires the package 'EQL' to work.`
