### 2019-08-06 #1566 @TrashBirdEcology
Added dependency issue, [issue #56](https://github.com/JonasMoss/kdensity/issues/56)
