### 2019-08-06 JonasMoss/kdensity#57 @trashbirdecology
[JOSS Review]: README.rmd minor comments/suggestions
JOSS Review [1566](https://github.com/openjournals/joss-reviews/issues/1566)

>"This is a good idea because the resulting estimator will be better than an ordinary kernel density estimator whenever the true density is close to your suggestion; and the estimator can be superior to the ordinary kernel density estimator even when the suggestion is pretty far off.">

Suggest editing this sentence by
- [x]  removing "this is a good idea"
- [x]  please explain what is meant by "better than an ordinary KDE"
------------------

The following is in the Section __Installation__ but is useage and not installation:

| Call the `library` function and use it just like `stats:density`, but with optional additional arguments.
```{r simpleuse, echo = TRUE, eval = FALSE}
library("kdensity")
plot(kdensity(mtcars$mpg, start = "normal"))
```

Clarify what the 'return value' is in:
>"Since the return value of `kdensity` is a function, it is callable, as in:"

---------------------
Revise the bold in sentence:
> "The following example __uses the data set plots both__ a gamma-kernel density estimate with a gamma start (black), the fully parametric gamma density (red), and an ordinary density estimate (blue)."

-------------------------
Consider editing for grammar in the parts in bold:
>If you encounter a bug, have a feature request or need some help, don’t hesitate to open an __issue. If__ you want to contribute, make a pull request. This project follows a Contributor Code of Conduct.
