### 2019-08-06 JonasMoss/kdensity#58 @trashbirdecology
JOSS REVIEW 1566: Minor spelling errors in documentation
### JOSS Review [1566](https://github.com/openjournals/joss-reviews/issues/1566)

Some _potential_ spelling issues and missing items in documentation found using spell check:

> devtools::spell_check()
DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
  WORD             FOUND IN
bw               add_bw.Rd:5,12,18
                 get_standard_bw.Rd:5,20
likelilood       README.md:129
                 README.Rmd:88
Miscellanea      kdensity.Rd:118
                 README.md:185
                 README.Rmd:125
na               kdensity.Rd:52
nrd              bandwidths.Rd:29
unneccessary     tutorial.Rmd:137
>
