### 2019-08-06 JonasMoss/kdensity#59 @trashbirdecology
JOSS Review 1566 Function kdensity.R review
### JOSS Review [1566](https://github.com/openjournals/joss-reviews/issues/1566)

In the function `kdensity.R`, you have a stop command if NAs are present in X. The line following the stop command removes NAs. So either replace 'stop' with 'warning' or 'message', or remove ` x = x[!is.na(x)] entirely:

```
if(has.na) {
    if(!na.rm) stop("x contains NAs and na.rm = FALSE.")
    x = x[!is.na(x)]
  }
```
---------------
 Consider removing the word "massage":
>  Now we massage and handle the combinations of kernel, start and support.
---------------
Consider adding error or message,  or a default value for @param: start, or remove NULL default for @param:start. If start=NULL, function will fail at line:
> kss_list = get_kernel_start_support(kernel, start, support)
