### 2019-08-07 4a515f5
:books: JOSS Review 1566 Function kdensity.R review
Removed "massage" from comments.
