+++ R/kdensity.R
@@ -149,7 +149,7 @@ kdensity = function(x, bw = NULL, adjust = 1, kernel = NULL, start = NULL,
   ## The case of bw == Inf is special! In this case, we return the parametric
   ## start itself.
 
-  ## Now we massage and handle the combinations of kernel, start and support.
+  ## Now we handle the combinations of kernel, start and support.
   ## This is fancy defaults management.
   kss_list = get_kernel_start_support(kernel, start, support)
 
