### 2019-08-07 JonasMoss/kdensity#59 @JonasMoss
The first issue is intended, as x = x[!is.na(x)] is never seen when the function is stopped, but is seen if it isn't. Take e.g.

```
f = function(x, na.rm = FALSE) {

  has.na = any(is.na(x))

  if(has.na) {
    if(!na.rm) stop("x contains NAs and na.rm = FALSE.")
    x = x[!is.na(x)]
  }

  x

}

> f(0, na.rm = TRUE)
[1] 0
> f(c(NA, 0), na.rm = TRUE)
[1] 0
> f(c(NA, 0), na.rm = FALSE)
Error in f(c(NA, 0), na.rm = FALSE) : x contains NAs and na.rm = FALSE.
```

I changed the code in 5747aa0 to make it clearer.

---
Removed "massage" in 4a515f52c63edc651b52daf5db22a71acfb98665.

---
Do you have an example of when the function fails? When I run it with only `NULL` it works:

```
> kdensity:::get_kernel_start_support(kernel = NULL, start = NULL, support = NULL)
$kernel
$kernel$kernel
function(y, x, h) stats::dnorm((y-x)/h)
<environment: namespace:kdensity>
...
```
