### 2019-08-07 934b0b9
[JOSS Review]: Installation instructions:
Added EQL to imports in description etc.
