### 2019-08-07 JonasMoss/kdensity#56 @JonasMoss
Added EQL as an import in 934b0b98ce9e31def79f30f3f30ab2fe5d5b0e0c.
