### 2019-08-07 b96ec08
[JOSS Review]: README.rmd minor comments/suggestions
