### 2019-08-07 JonasMoss/kdensity#57 @JonasMoss
Thanks! Updated in b96ec08bdb9a90f3364b7ecf090ab69ee8b13d0c.
