### 2019-08-07 f685e1c
JOSS REVIEW 1566: Minor spelling errors in documentation
Fixed spelling errors.
