### 2019-08-07 JonasMoss/kdensity#58 @JonasMoss
Fixed in f685e1c.
