### 2019-08-07 5747aa0
JOSS Review 1566 Function kdensity.R review
Changed if to assertthat.
