### 2019-08-07 #1566 @JonasMoss
Thanks a lot @TrashBirdEcology ! I've addressed your issues with details on the repository's issues page.
