### 2019-09-03 JonasMoss/kdensity#48 @gvegayon
Works for me! Thanks
