### 2019-09-03 JonasMoss/kdensity#51 @gvegayon
That's nice. Thanks. You need to fix one of the links (see note on commit).
