### 2019-09-03 #1566 @gvegayon
OK @JonasMoss, the changes look good. Only one minor issue regarding a broken link. Thanks.
