### 2019-10-01 015f9a8
Update NEWS.md
