+++ NEWS.md
@@ -1,3 +1,9 @@
+## kdensity 1.1.0 (2019-01-10)
+
+* Journal of Open Source Software Release.
+* Safer kdensity object structure.
+* Cleaned up readme and documentation.
+
 ## kdensity 1.0.1 (2019-07-11)
 
 * Fixed the `laplace` kernel.
