### 2019-10-01 #1566 @JonasMoss
Thanks! Here's the doi of the Zenodo archive.
10.5281/zenodo.3466547
https://zenodo.org/record/3466547
