### 2019-10-01 #1566 @terrytangyuan
@JonasMoss Could you take a look at your recent changes to see what caused the above compilation failure?
