### 2019-10-01 c20afd5
Changed encoding to UTF-8.
