### 2019-10-01 #1566 @JonasMoss
I changed the encoding from ANSI to UTF-8 at the Github repo now.
