### 2019-10-01 bcfdbd3
Changed ANSI to UTF-8.
