+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@article{hjort_glad_1995,
+ï»¿@article{hjort_glad_1995,
   title={Nonparametric density estimation with a parametric start},
   author={Hjort, Nils Lid and Glad, Ingrid K},
   journal={The Annals of Statistics},
@@ -121,7 +121,7 @@
 
 @Manual{Ake,
     title = {Ake: Associated Kernel Estimations},
-    author = {W. E. Wansouwé and S. M. Somé and C. C. Kokonendji},
+    author = {W. E. WansouwÃ© and S. M. SomÃ© and C. C. Kokonendji},
     year = {2015},
     note = {R package version 1.0},
     url = {https://CRAN.R-project.org/package=Ake},
