### 2019-10-01 #1566 @terrytangyuan
@JonasMoss Could you fix the invalid DOI shown above?
