+++ paper/paper.bib
@@ -61,7 +61,7 @@
   pages={653--671},
   year={1994},
   publisher={Wiley Online Library},
-  doi = {doi:10.2307/2346189}
+  doi = {doi:10.1111/j.2517-6161.1994.tb02006.x}
 }
 
 @Manual{r,
