### 2019-10-01 #1566 @terrytangyuan
@openjournals/joss-eics Could you take from here? The paper looks good to me. I manually verified that the DOI has been fixed.
