### 2019-10-01 #1566 @kyleniemeyer
Hi @JonasMoss, not sure if this is causing the whedon error (and if so, why it's only showing up for one of them), but if you look in the list of references, a number of them have repeated "doi:" in the DOIs.

Please remove the extra `doi:` from these references: Chen 1999, Chen 2000, Hjort 1995, Jones 2007, Marron 1994, Silverman 2018 (and any others, if I missed any)

(in other words, change `doi = {doi:10.1214/aos/1176324627}` to just `doi = {10.1214/aos/1176324627}` (e.g.) in the references where applicable.)
