### 2019-10-01 49a4f86
Cleaning up paper.bib.
