+++ paper/paper.bib
@@ -1,128 +1,127 @@
 ﻿@article{hjort_glad_1995,
-  title={Nonparametric density estimation with a parametric start},
-  author={Hjort, Nils Lid and Glad, Ingrid K},
-  journal={The Annals of Statistics},
-  pages={882--904},
-  year={1995},
-  publisher={JSTOR},
-  doi = {doi:10.1214/aos/1176324627}
+	title={Nonparametric density estimation with a parametric start},
+	author={Hjort, Nils Lid and Glad, Ingrid K},
+	journal={The Annals of Statistics},
+	pages={882--904},
+	year={1995},
+	doi={10.1214/aos/1176324627}
 }
 
 @article{jones2007miscellanea,
-  title={Kernel-type density estimation on the unit interval},
-  author={Jones, MC and Henderson, DA},
-  journal={Biometrika},
-  volume={94},
-  number={4},
-  pages={977--984},
-  year={2007},
-  publisher={Oxford University Press},
-  doi = {doi:10.1093/biomet/asm068}
+	title={Kernel-type density estimation on the unit interval},
+	author={Jones, MC and Henderson, DA},
+	journal={Biometrika},
+	volume={94},
+	number={4},
+	pages={977--984},
+	year={2007},
+	publisher={Oxford University Press},
+	doi={10.1093/biomet/asm068}
 }
 
 @article{chen2000probability,
-  title={Probability density function estimation using gamma kernels},
-  author={Chen, Song Xi},
-  journal={Annals of the Institute of Statistical Mathematics},
-  volume={52},
-  number={3},
-  pages={471--480},
-  year={2000},
-  publisher={Springer},
-  doi = {doi:10.1023/A:1004165218295}
+	title={Probability density function estimation using gamma kernels},
+	author={Chen, Song Xi},
+	journal={Annals of the Institute of Statistical Mathematics},
+	volume={52},
+	number={3},
+	pages={471--480},
+	year={2000},
+	publisher={Springer},
+	doi={10.1023/A:1004165218295}
 }
 
 @article{chen1999beta,
-  title={Beta kernel estimators for density functions},
-  author={Chen, Song Xi},
-  journal={Computational Statistics \& Data Analysis},
-  volume={31},
-  number={2},
-  pages={131--145},
-  year={1999},
-  publisher={Elsevier},
-  doi = {doi:10.1016/S0167-9473(99)00010-9}
+	title={Beta kernel estimators for density functions},
+	author={Chen, Song Xi},
+	journal={Computational Statistics \& Data Analysis},
+	volume={31},
+	number={2},
+	pages={131--145},
+	year={1999},
+	publisher={Elsevier},
+	doi={10.1016/S0167-9473(99)00010-9}
 }
 
 @book{silverman2018density,
-  title={Density estimation for statistics and data analysis},
-  author={Silverman, Bernard W},
-  year={2018},
-  publisher={Routledge},
-  doi = {doi:10.1201/9781315140919}
+	title={Density estimation for statistics and data analysis},
+	author={Silverman, Bernard W},
+	year={2018},
+	publisher={Routledge},
+	doi={10.1201/9781315140919}
 }
 
 @article{marron1994transformations,
-  title={Transformations to reduce boundary bias in kernel density estimation},
-  author={Marron, James Stephen and Ruppert, David},
-  journal={Journal of the Royal Statistical Society: Series B (Methodological)},
-  volume={56},
-  number={4},
-  pages={653--671},
-  year={1994},
-  publisher={Wiley Online Library},
-  doi = {doi:10.1111/j.2517-6161.1994.tb02006.x}
[…]
