### 2019-10-01 #1566 @JonasMoss
@kyleniemeyer OK, done it.

The doi 10.2307/2346189 was in fact invalid acccording to doi.org though. I replaced it with the correct doi.

The string "10.2307/2346189" isn't in the paper.bib on the Github repo (i.e. search finds nothing), so it looks like whedon isn't looking at the right place. Perhaps the document must be compiled again?
