### 2019-10-01 #1566 @kyleniemeyer
@JonasMoss nearly there! I just have one final paper edit: could you add a sentence or two at the beginning of the paper to satisfy this requirement:
> A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.

(from our list of things that JOSS papers [should contain](https://openjournals.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain))
