### 2019-10-03 c7d6525
Update paper.md
