+++ paper/paper.md
@@ -23,33 +23,34 @@ bibliography: paper.bib
 ---
 
 # Summary
+Estimating probability densities is an important task in many data science problems.
+The R package [@r] `kdensity` is an alternative --- and in some cases, more correct ---
+way of performing non-parametric density estimation than the standard `density` function
+in R.
+A primary reason for this is that `kdensity` allows one to 
+specify the support of the density estimate through an initial parametric density estimate.
 
-Kernel density estimation [@silverman2018density] is a popular method for 
-non-parametric density estimation based on placing kernels on each data point. 
+The `density` function is based on classical kernel density estimation [@silverman2018density],
+where symmetric kernels are placed on each data point. 
 @hjort_glad_1995 extended kernel density estimation with *parametric starts*.
 The parametric start is a parametric density that is multiplied with the kernel
 estimate. When the data-generating density is reasonably close to the parametric
 start density, kernel density estimation with that parametric start will outperform
 ordinary kernel density estimation.
 
-Asymmetric kernels are useful for estimating densities on the half-open interval $\left[0,\infty\right)$ and bounded intervals such as $\left[0, 1\right]$. On 
-such intervals symmetric kernels are prone to serious boundary bias that should
-be corrected [@marron1994transformations]. Asymmetric kernels are designed to
-avoid boundary bias.
+Moreover, when estimating densities on the half-open interval $\left[0,\infty\right)$ and bounded intervals, such as $\left[0, 1\right]$, symmetric kernels are prone to serious boundary bias that should be corrected [@marron1994transformations]. 
+Asymmetric kernels have been designed to avoid boundary bias and many of them are implemented in `kdensity` in addition to the classical symmetric kernels.
+For the unit interval, the Gaussian copula kernel of @jones2007miscellanea and the beta kernels of @chen1999beta are supported. 
+The gamma kernel of @chen2000probability is available for the half-open interval.
 
-`kdensity` is an R package [@r] to calculate and display kernel density 
-estimates using non-parametric starts and potentially asymmetric kernels. In 
-addition to the classical symmetric kernels, `kdensity` supports the following 
-asymmetric kernels: For the unit interval, the Gaussian copula kernel of @jones2007miscellanea and the beta kernels of @chen1999beta are supported. On 
-the half-open interval the gamma kernel of @chen2000probability is supported. 
 The supported non-parametric starts include the normal, Laplace, Gumbel, 
 exponential, gamma, log-normal, inverse Gaussian, Weibull, Beta, and Kumaraswamy
 densities. The parameters of all parametric starts are estimated using maximum 
 likelihood. The implemented bandwidth selectors are the classical bandwidth 
 selectors from `stats`, unbiased cross-validation, the Hermite polynomial method 
 from @hjort_glad_1995, and the tailored bandwidth selector for the Gaussian 
-copula method of @jones2007miscellanea. User defined parametric starts, 
-kernels and bandwidth selectors are also supported. 
+copula method of @jones2007miscellanea. User-defined parametric starts, 
+kernels and bandwidth selectors can also be set. 
 
 Several `R` packages deal with kernel estimation, see @deng2011 for an overview. 
 While no other `R` package handles density estimation with parametric starts, 
