### 2019-10-03 #1566 @JonasMoss
@kyleniemeyer OK! We added two sentences to the beginning of the paper now.
