### 2019-10-03 #1566 @kyleniemeyer
@JonasMoss thanks! just one minor correction: I think you are missing "R" at near the beginning of the second sentence: "This (R Core Team, 2019) package..." -> "This R (R Core Team, 2019) package..."
